posix: completed signal.h, added chdir and getppid

+++ src/sysindep/signal/sigprocmask.c
@@ -4,5 +4,5 @@
 
 int sigprocmask(int how, const sigset_t *set, sigset_t *oldset)
 {
-    return syscall(SYS_sigprocmask, how, (long)set, (long)oldset);
+    return (int)syscall(SYS_sigprocmask, how, (long)set, (long)oldset);
 }
